+++ src/impl/xliff-file.spec.ts
@@ -34,6 +34,7 @@ describe('ngx-i18nsupport-lib xliff 1.2 test spec', () => {
         let ID_WITH_TWO_SOURCEREFS = '78eab955529ba0f1817c84991d9175f55bfdf937'; // an ID with 2 source elements
         let ID_WITH_TAGS = '7e8dd1fd1c57afafc38550ce80b5bcc1ced49f85';
         let ID_WITH_TAGS_2 = '7e8dd1fd1c57afafc38550ce80b5bcc1ced49f85xxx'; // same with </x> tags
+        let ID_UNTRANSLATED_DESCRIPTION = 'a52ba049c16778bdb2e5a19a41acaadf87b104dc';
 
         it('should read xlf file', () => {
             const file: ITranslationMessagesFile = readFile(MASTER1SRC);
@@ -155,5 +156,64 @@ describe('ngx-i18nsupport-lib xliff 1.2 test spec', () => {
             expect(tu2).toBeFalsy(); // should not exist any more
         });
 
+        it ('should translate source without or with target', () => {
+            const file: ITranslationMessagesFile = readFile(MASTER1SRC);
+            const tu: ITransUnit = file.transUnitWithId(ID_TRANSLATED_SCHLIESSEN);
+            expect(tu).toBeTruthy();
+            expect(tu.targetContent()).toBeFalsy();
+            let isDefaultLang: boolean = false;
+            let copyContent: boolean = true;
+            // first translate
+            tu.translate('Anwendung läuft');
+            const file2: ITranslationMessagesFile = TranslationMessagesFileFactory.fromUnknownFormatFileContent(file.editedContent(), null, null);
+            const tu2: ITransUnit = file2.transUnitWithId(ID_TRANSLATED_SCHLIESSEN);
+            expect(tu2.targetContentNormalized()).toBe('Anwendung läuft');
+            expect(tu2.targetState()).toBe('final');
+            // translate again
+            tu2.translate('Anwendung funktioniert');
+            const file3: ITranslationMessagesFile = TranslationMessagesFileFactory.fromUnknownFormatFileContent(file2.editedContent(), null, null);
+            const tu3: ITransUnit = file3.transUnitWithId(ID_TRANSLATED_SCHLIESSEN);
+            expect(tu3.targetContentNormalized()).toBe('Anwendung funktioniert');
+            expect(tu3.targetState()).toBe('final');
+        });
+
+        it ('should copy source to target for default lang', () => {
+            const file: ITranslationMessagesFile = readFile(MASTER1SRC);
+            const tu: ITransUnit = file.transUnitWithId(ID_UNTRANSLATED_DESCRIPTION);
+            expect(tu).toBeTruthy();
+            expect(tu.targetContent()).toBeFalsy();
+            let isDefaultLang: boolean = true;
+            let copyContent: boolean = false;
+            tu.useSourceAsTarget(isDefaultLang, copyContent);
+            const file2: ITranslationMessagesFile = TranslationMessagesFileFactory.fromUnknownFormatFileContent(file.editedContent(), null, null);
+            const tu2: ITransUnit = file2.transUnitWithId(ID_UNTRANSLATED_DESCRIPTION);
+            expect(tu2.targetContentNormalized()).toBe('Beschreibung zu {{0}} ({{1}})');
+        });
+
+        it ('should copy source to target for non default lang if wanted', () => {
+            const file: ITranslationMessagesFile = readFile(MASTER1SRC);
+            const tu: ITransUnit = file.transUnitWithId(ID_UNTRANSLATED_DESCRIPTION);
+            expect(tu).toBeTruthy();
+            expect(tu.targetContent()).toBeFalsy();
+            let isDefaultLang: boolean = false;
+            let copyContent: boolean = true;
+            tu.useSourceAsTarget(isDefaultLang, copyContent);
+            const file2: ITranslationMessagesFile = TranslationMessagesFileFactory.fromUnknownFormatFileContent(file.editedContent(), null, null);
+            const tu2: ITransUnit = file2.transUnitWithId(ID_UNTRANSLATED_DESCRIPTION);
+            expect(tu2.targetContentNormalized()).toBe('Beschreibung zu {{0}} ({{1}})');
+        });
+
+        it ('should not copy source to target for non default lang if not wanted', () => {
+            const file: ITranslationMessagesFile = readFile(MASTER1SRC);
+            const tu: ITransUnit = file.transUnitWithId(ID_UNTRANSLATED_DESCRIPTION);
+            expect(tu).toBeTruthy();
+            expect(tu.targetContent()).toBeFalsy();
+            let isDefaultLang: boolean = false;
+            let copyContent: boolean = false;
+            tu.useSourceAsTarget(isDefaultLang, copyContent);
+            const file2: ITranslationMessagesFile = TranslationMessagesFileFactory.fromUnknownFormatFileContent(file.editedContent(), null, null);
+            const tu2: ITransUnit = file2.transUnitWithId(ID_UNTRANSLATED_DESCRIPTION);
+            expect(tu2.targetContent()).toBeFalsy();
+        });
     });
 });